Introduce Rate and Ratio opaque types for interest rates and shares
Add Rate opaque type for interest rates, yields, inflation, and returns
(~22 case class fields). Add Ratio opaque type for shares, probabilities,
and fractions (~25 fields). Extend types.scala with >=, <=, /, max, min
operations. Update all ~500 call sites across 22 source files and 54 test
files with .toDouble at read boundaries and Rate()/Ratio() at write
boundaries. All 1250 tests pass.

+++ src/test/scala/sfc/agents/BankingSectorPropertySpec.scala
@@ -22,7 +22,7 @@ class BankingSectorPropertySpec extends AnyFlatSpec with Matchers with ScalaChec
 
   "clearInterbank" should "always produce interbankNet that sums to zero" in {
     forAll(genBankingSectorState) { (bs: BankingSectorState) =>
-      val cleared = BankingSector.clearInterbank(bs.banks, bs.configs, bs.interbankRate)
+      val cleared = BankingSector.clearInterbank(bs.banks, bs.configs, bs.interbankRate.toDouble)
       val netSum = cleared.map(_.interbankNet.toDouble).sum
       netSum shouldBe 0.0 +- 1.0  // tolerance for floating-point
     }
@@ -114,7 +114,7 @@ class BankingSectorPropertySpec extends AnyFlatSpec with Matchers with ScalaChec
 
   "clearInterbank" should "keep reserves non-negative" in {
     forAll(genBankingSectorState) { (bs: BankingSectorState) =>
-      val cleared = BankingSector.clearInterbank(bs.banks, bs.configs, bs.interbankRate)
+      val cleared = BankingSector.clearInterbank(bs.banks, bs.configs, bs.interbankRate.toDouble)
       cleared.foreach(_.reservesAtNbp.toDouble should be >= 0.0)
     }
   }
